修正理论NP率计算时有概率会产生精度错误的问题. 更改部分计算变量从double修改为decimal. 优化代码.

+++ Properties/GlobalPathsAndDatas.cs
@@ -61,10 +61,10 @@ namespace Altera.Properties
         public static string CurveType;
         public static string svtIDListStr;
         public static string IDListStr;
-        public static double notrealnprate;
+        public static decimal notrealnprate;
         public static int LvExpCurveLvCount;
         public static int[] CurveBaseData;
-        public static double ymax;
+        public static decimal ymax;
         public static int classid;
         public static string AS1N;
         public static string AS1D;
